requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
openai>=1.0.0
python-docx>=1.1.0
lxml>=5.0.0
